Removed hard coded ROM offsets, moved to file - enable viewing of all release ROMs Moved functionality from MainFrame to Palette and Sprite - these classes now automatically load in the relevant ROM data Changed sprite drawing to be centred on origin, rather than cropped. This means that sprite image positioning is preserved.

+++ landstalker_gfx/ImageBuffer.h
@@ -18,7 +18,7 @@ public:
 	ImageBuffer(size_t width, size_t height);
 	void Clear();
 	void Resize(size_t width, size_t height);
-	void InsertTile(size_t x, size_t y, uint8_t palette_index, const Tile& tile, const Tileset& tileset);
+	void InsertTile(int x, int y, uint8_t palette_index, const Tile& tile, const Tileset& tileset);
 	bool WritePNG(const std::string& filename, const std::vector<Palette>& pals);
 	void InsertBlock(size_t x, size_t y, uint8_t palette_index, const BigTile& block, const Tileset& tileset);
 	const std::vector<uint8_t>& GetRGB(const std::vector<Palette>& pals) const;
